Extract Yearn vault deployment function
